backend/requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.0
Requests==2.31.0
